refactor: add unified components barrel export
Add src/components/index.ts for unified imports:

// Before - multiple import paths
import { Button } from '@/components/ui';
import { Layout } from '@/components/layout';
import { useToast } from '@/components/feedback/Toast';
import { Modal } from '@/components/feedback/Modal';

// After - single unified import
import { Button, Layout, useToast, Modal } from '@/components';

Structure:
src/components/
├── index.ts         # Unified barrel export (NEW)
├── ui/              # Button, Input, Card, etc.
├── layout/          # Layout, PageLoadingFallback
├── feedback/        # Toast, Modal, SaveIndicator
├── auth/            # ProtectedRoute, AuthFormLayout
└── financial/       # FinancialMetrics, Charts

Benefits:
- Single import path for all components
- Easier to use and remember
- Better maintainability
- Cleaner code

Build Status: ✅ Successful
