+++ src/components/financial/FinancialMetrics.tsx
@@ -12,7 +12,7 @@
 
 import React from 'react';
 import { DollarSign, TrendingUp, Target, AlertCircle, CheckCircle2 } from 'lucide-react';
-import { Badge } from '@/components/ui';
+import { Badge } from '@/components';
 import { formatCurrency, formatNumber, cn } from '@/common/utils';
 
 export interface FinancialMetricsData {
