+++ src/components/index.ts
@@ -1,0 +1,36 @@
+/**
+ * 파일명: index.ts
+ * 
+ * 파일 용도:
+ * 컴포넌트 통합 barrel export
+ * 모든 컴포넌트를 @/components에서 import 가능
+ * 
+ * 사용법:
+ * import { Button, Layout, ToastProvider, Modal } from '@/components';
+ */
+
+// ============================================
+// UI 프리미티브 (Button, Input, Card 등)
+// ============================================
+export * from './ui';
+
+// ============================================
+// 레이아웃 (Layout, Loading Fallback)
+// ============================================
+export * from './layout';
+
+// ============================================
+// 피드백 (Toast, Modal, SaveIndicator)
+// ============================================
+export * from './feedback';
+
+// ============================================
+// 인증 (AuthFormLayout, PasswordInput, ProtectedRoute)
+// ============================================
+export * from './auth';
+
+// ============================================
+// 재무 (FinancialMetrics, FinancialCharts)
+// ============================================
+export * from './financial';
+
