+++ src/hooks/useExportDocument.ts
@@ -13,7 +13,7 @@
  */
 
 import { useState, useCallback } from 'react';
-import { useToast } from '@/components/feedback/Toast';
+import { useToast } from '@/components';
 import { useProjectStore } from '@/stores/useProjectStore';
 import { 
   ExportFormat,
