+++ src/pages/calculator/CTABanner.tsx
@@ -13,7 +13,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Sparkles, ArrowRight, Lock, Zap, Shield, Save } from 'lucide-react';
-import { Button } from '@/components/ui';
+import { Button } from '@/components';
 import { cn } from '@/common/utils';
 
 interface CTABannerProps {
